Determine accuracy on 30/70 split
This code reads in the data from concat_roi_avg.m and
splits it into 70% train and 30% test. It normalizes all
the data to speed up glasso, generates 2 precision matricies (one
for each class) and then determines which class is more likely to
be the right class for each test item. Then we calculate accuracy